Update doc for 21.12 release (#58)
* Update doc for 21.12 release


* update for 21.12 release


* update


* Update

+++ examples/pca/main.scala
@@ -26,7 +26,7 @@ val dataDf = spark.createDataFrame(
   (0 until rows).map(_ => Tuple1(Array.fill(dim)(r.nextDouble)))).withColumnRenamed("_1", "feature")
 
 // use udf to meet ML algo input requirement: Vector input
-val convertToVector = udf((array: Seq[Float]) => {
+val convertToVector = udf((array: Seq[Double]) => {
   Vectors.dense(array.map(_.toDouble).toArray)
 })
